(Fix) Icon issues with sortable columns
This fixes a issue with sortable columns like on the `my active table`
page.

Unfortunately, due to the order that config files are loaded, the
`sortablecolumns.php` does NOT have access to the `other.php`
configuration values. So until we come up with a better solution we
opted to hard code the `FontAwesome Type` in `sortablecolumns.php`