[BUG FIX] Bonus Exchange
Fixes a bug where the bonus exchange route, is callable via a GET method, and can be maliciously executed when placing it within a image bbcode tag.

This fix shifts the route to a post method, then also uses csrf to allow for some mitigation of this bug, till a more appropriate fix can be put in place.

Co-Authored-By: HDVinnie <unit3d@protonmail.com>