(Update) Request Modals View
Update routes reference (coming).

Remove unneeded hidden variable.

TODO: Fix this ugly naming of torrentRequest ... shouldn't torrent_request be more appropriate?

Or maybe request as root since we are giving requests relationships.

So it would be safe to say

request->id because we now have a relationship to tie it to.