(Fix) search result renders empty element
Resubmitted the change again.
When the $page posted back, it maybe 0. In order to use it for array index, we need to avoid negative value after deducting it by 1.
That 0 case occurs when the searching takes place at default Group View loading. On the other hand, if the searching combined with a click on Paginator, the Paginator already posts a correct $page value instead of 0.