fix: Move calls to env() out of UsersTableSeeder
env() calls should be made *only* from within config files.

Note that the 'owner-username' is not prepended with 'default-' because,
unlike the other two values, it should not be changed after
installation, specifically because the account should always be
looked-up by username and not ID (despite existing code that currently
does the latter).