Update: Stops users from passively seeding to themselves.
This stops announce from sending peers to the users client that are from the same user.
This does not stop a user from manually adding their other peer to the client to seed to themselves.
With that being said if a users is found to be seeding to themselves that means they are activly trying to break the rules and should be banned.