update: logging config
In a non-local environment, it is recommended that your log level should not be set to debug because of various cost factors:

Cost of Storage: Debug logs are the most verbose and will increase the cost of storing logs.
Cost of Processing: If you use a third party service to accumulate logs across your servers, there will be incremental costs for processing logs based on your service provider.