Switched uniqid for random_bytes.
uniqid isn't cryptographically secure, it is a pseudorandom number generator meaning that its values can be deduced or easily guessed within a computationally reasonable amount of guesses, this branch removes its security-oriented implementations and uses random_bytes instead.

Note: When a CRNG is not required, uniqid and other functions similar to it (e.g mt_rand & rand) are fine to use as PRNGs which is why I've left it in use for the filename-generation system.