Add "No torrent found" message
On a fresh install of the system, this command prints a log entry with an error every time it's executed. This creates a temporary model (new instead of create) that can be passed to the helper functions. This is a quick and dirty fix which should be thought about a little bit more, but it does its job for now.