fix: floating labels relying on placeholder="" in chrome
Form input labels rely on the :place-holder-shown css property to float the label whenever the placeholder disappears (i.e. any time there's text inside) that way a user can click outside of the form input and the label will stay floating. Unfortunately, it seems that because only `placeholder=""` is used with the placeholder attribute being empty, Chrome interprets this as no place-holder being shown even if no text has been entered in the form input, triggering the label to float automatically.
To fix this, we must add a space to every placeholder attribute (`placeholder=" "`)