fix: featured buttons and animation
The code for having the carousel rotate was unintentionally commented out.

Additionally, there seems to be a rounding error somewhere when there are 6 or more featured torrents where scrolling one pixel past halfway through the next card isn't enough to get it to scroll the rest of the way on its own using css scroll behavior. So that's now been updated to 2 pixels which seems to work with 6 featured torrents.

Finally, the card gap (currently set to 16 px) wasn't accounted for in the wrap around for when the last card is reached. Now it is.