fix: bad mime detection
Many users often remove the announce url from uploaded torrents (since other tracker codebases are known to log urls of uploaded torrents to breach user accounts on other sites). This unfortunately causes the mime detection to sometimes (not all the time) fail preventing the user from uploading. Now that we're validating by file extension, it should still prevent the case of users accidentally uploading an unrelated file which was the main issue to be solved.