update: use lower level data structures for announce
We can't pass Models or collections into a job, as Laravel will apply their own custom serialization format that errors when trying to deserialize. So, we now use basic arrays. Logic relying on Eloquent Collections is now done manually. Reducing the extra overhead might be good for performance too. Also combined the case statement as many of the steps were the same and they would become much longer with the extra code required.