fix: history upsert seedtime
It turns out Laravel will sort the keys in the upsert array alphabetically. This breaks the logic for seedtime calculation which ensures to only increase the seedtime if the previous announce had a `seeder` value of 1. To fix this, we include the sql code necessary to update the `seeder` column inside the `DB::raw()` statement of the `seedtime` update, to make sure `seeder` is updated after `seedtime`. A hacky solution, but seems to be the best way to go about it for now.