update: inline announce job into announce
This prevent duplicate queries from being made since laravel jobs by default refetch Models from the database if they are an argument. The main issue with doing this before was that the Peer/History inserts/updates are expensive, but we now only do an O(1) insert into a redis queue and upsert them all at once later. User and Torrent updates are occasional enough that they shouldn't be an issue, and PersonalFreeleech and FreeleechToken queries should have small enough tables and be cached fast enough that there shouldn't be any issues either.