update: increase announce cache expiry
Cache busting seems to be working enough that we probably don't need this fail-safe, but we can bump it up a bit at least. The key and value (<50 bytes for non-existant values) are small enough that we don't have to worry about running out of memory with the current rate limit in case of a few malicious users.

Co-authored-by: HDVinnie <hdinnovations@protonmail.com>