fix: redundant movie db query for tv api calls
Tv torrents don't have the `movie` relation eager loaded so it will try and fetch the `movie` relation if it's not eager loaded. Since we don't care about movies on a tv torrent, use a variable that doesn't have the same name as a relation so that it can't get loaded.