fix: json serialization in database migrations
When the switch was made in the database migrations to use the query builder instead of eloquent models in e67f7dc01387daa4f00e44eeb8ffa9322f756c4a, the json casting provided by the eloquent models weren't accounted for. We have to deserialize and serialize the json manually now.