fix: torrent file hierarchy dropdown expansion indicator
Without this, the dropdown will have two triangles that show the state of if the dropdown is open or not. We do this by using the default list style marker instead of custom font awesome icons. Because of this, we can remove the font-awesome icon code and make sure that when the depth is 0, the extra 20px of margin is kept. I switched out the padding for grid gap while I was at it and used blade @styles for the conditional logic.