fix: autocommit on pull requests from forks
The pull_request event runs in the context of the PR's repo, while the push event runs in the context of the current repo. So, we need to order the events so that if a user submits a PR from their fork, that we run in the context of the fork instead of the current repo (which would result in the branch not being found).