fix: hierarchy file sorting
A bit of a hack is needed to sort files before directories located in the same directory. For example, if a torrent has two files: `BDMV/STREAM/00000.m2ts` and `BDMV/index.bdmv`, users will expect `BDMV/STREAM/` to be ordered before `BDMV/index.bdmv` in hiearchical sorting/grouping. In order to do this, we append a slash and the highest value of the ascii table (`chr(255)`) to the directory of each filepath, so that we e.g. sort `MOVIE_NAME/STREAM/<0xFF>` before `MOVIE_NAME/<0xFF>`, since `0xFF` will always (assuming sane filepaths...) come after `S`. Previously, we used the tilde for this hack, which has an ascii code of 126 which comes after mostly all characters, but let's max it out this time around to `chr(255)`. Also, there was an issue where a file located in the root directory would have `.` returned as its `dirname()`. `.` comes before all letters in the ascii table which breaks sorting when a torrent had both a directory and a file in the root directory, such as there being a `disc.inf` file in the root directory next to the `MOVIE_NAME` directory (In that case, the `disc.inf` file would be sorted before the `MOVIE_NAME` directory).