update: torrent group monthly season pack naming
For shows that are packed by month, with a torrent title following the format `TV Show Name 2024-05 1080p H.264`, the year is saved as the season and there is not way to disambiguate between the months within. So instead of the row saying `1080p H.264` with no context as to which month it is, it becomes `05 1080p H.264`. We also now match between all years between the first air date and the last end date plus/minus 1 instead of matching any date format within the title.