update: force redirect `/index.php/` requests
This was never intended, and was recently discovered when an old site was migrated to UNIT3D and its users had browser bookmarks to https://domain.tld/index.php. Users browsing from this url have all their links generated with the `index.php` included, include the announce urls in their torrents, which isn't great and causes issues with unit3d-announce.