fix: register page seo
When someone follows the link without the invite code when the site is configured as invite only, we show an error message saying open registration is disabled. Unfortunately, if the code is missing, we redirect the user with a `?code=null` query parameter instead of omitting the code, which prevents the error message on the registration page from being shown. Let's fix this by omitting the query parameter if a code wasn't provided.