update: save prewarned timestamp
Allows users to more easily see when they were prewarned in case they no longer have their notification.

It was decided to discard previous prewarn values and have the system prewarn users again if applicable.

It was decided that keeping the history of prewarn values wasn't valuable for this migration.