fix: double url encoding in bbcode comparison
The url is already sanitized and ran through `htmlspecialchars` at the end of the `sanitizeUrl` function. When using `{{ }}`, laravel runs the url through `htmlspecialchars` a second time, which breaks the url. Since the url is already properly sanitized, we can safely display the data without escaping in blade a second time.