update: add indexes to torrents table
Index on the name makes it so that the query to check if a torrent of the same name already exists upons upload is instant instead of taking 600 ms. Index on user_id, status, anon, deleted_at makes home page load instantly instead of taking 10+ seconds to calculate user stats.