fix: grouping name trimming regex again
We can't match on only the name because it doesn't trim `AKA`s. Let's try a lookahead this time to still search for the SxxEyy we searched for previously, but not trim it. Also fixes issue when media name would include slash which wasn't escaped when put into the regex. Also fixes issue with not matching complete season packs with season ranges.