fix: only delete audit log entries newer than unix epoch
It's easy for the config entry to recycle audits after a huge number of days. For example, a number of days less than the unix epoch once subtracted from the current time. Such a number will generate errors in the cron job audit recycling query. So, limit queries to use a datetime newer than the unix epoch to compensate.