update: improve footer performance
This function was ran in every view. The problem is for every partial, it was ran again. On pages like /mediahub/persons/x, this function was called 4000+ times because of all the partials, causing redis cache hits to use up more than 75% of the processing time to return the response. Fix by only calling it in the footer explicitly.