fix: broken api compatbility when no torrents found
There was always a null check here to return 404 when no torrents are found. But this null check was always comparing against a paginator instance... which would never be true. So when I changed it to an isEmpty when I changed the query from a database query to a meilisearch query, and thought that was the correct thing to do to make sure a 404 is still returned, that was incorrect, because a 404 was never returned in the first place. An empty array is proper here.