fix: conversation and ticket search tabs flashing
It was likely this was a bad copy-paste error from the person-credit tabs
which are by default hidden and are unhidden for tabs with greater than 0
items. This is not needed for ticket search and conversation search and adds
an unneeded flash of unstyled content. Remove `x-cloak` removes the flash.