fix: receiving chat pms from other users while in chat pm
When conversing in a chat pm and you receive a chat pm from another user, their message will appear inside of the current chat pm you have open.

This is because of incorrect logic inside Chatbox.vue. When user A sends a message to user B via the Chatter event, the message will have a subproperty `user` with user A's id. But user A is user B's target. So, these two properties need to be compared to determine if the message should be saved or not.