fix: validation on updating group autogroup
When `autogroup` was false, blade doesn't render anything (empty string). By rendering it as `false` using the `Js` facade, the value of `autogroup` is set. The value is synced with the checkbox in the form. When it was blank, the checkbox value was errored. With the error fixed, the correct value is submitted and the validation passes. Fixes #4401