refactor: move base view layout database queries inside view composers
Keeps code more organized by not having database queries inside the view. Just learned that these were a thing, and realized in the process why there was the issue with #4198 - it was because it was matching on `*`, i.e. all views, so it would render once for every single view, including each include.