update: change meilisearch proximity precision to `byAttribute`
This setting is more useful for full text search inside paragraphs of content and we already have it at the bottom of our ranking rules. The docs (https://www.meilisearch.com/docs/reference/api/settings#proximity-precision) mention a significant indexing performance boost when this is switched to `byAttribute` so let's take advantage of that.

Requires running `php artisan scout:sync-index-settings`.