update: remove phpstan nullsafe.neverNull rule
Imo, this is a bad rule. Although it is correct in that ?? will shortcircuit any null checks, I strongly prefer it being explicit as to which properties in the chain are nullable and which aren't. In fact, I wish we could enforce explit ?-> over -> when the property is nullable so that it's obvious which properties are null or not.