fix: notification comment tag tests
Tagging a user while creating a comment will both create a notification for the tag, as well as create a notification to the user who made the torrent/request. We only want to assert that the count of notifications created during the test is 1 only for the tag notifications and not count the other notification created.